Added IntelTracker to "made with" section (#1002)
IntelTracker is a project intended for players and speed-runners of No One Lives Forever (2000, PC). It tracks what intelligence items are found in the game as the player finds them. It uses ``electron-vue`` to allow it to run locally as an application, with ``vue`` handling UI updates and customization, and save-file watching using ``chokidar``.